HeroSection tests, render secondary text
