+++ components/HeroSection/HeroSection.module.css
@@ -18,12 +18,17 @@
 }
 
 .mainText {
-  font-size: 2rem;
   text-transform: uppercase;
   color: var(--mantine-color-white);
   
   @media (max-width: 600px) {
     font-size: 1.5rem;
+  }
+}
+
+.secondaryText {
+  @media (max-width: 600px) {
+    font-size: 1rem;
   }
 }
